requirements.txt
torch==2.7.0
python-dotenv
openai==1.75.0
transformers==4.51.3
datasets==3.5.0
predibase==2025.3.2
huggingface-hub==0.30.2
ipywidgets==8.1.6
matplotlib==3.10.1
peft==0.15.2